requirements.txt
# 生产依赖（本地开发完整安装）
crepe>=0.0.16
librosa>=0.10.0
pretty_midi>=0.2.10
music21>=9.1.0
torch>=2.0.0
gradio>=4.0.0
pyfluidsynth>=1.3.0
matplotlib>=3.7.0
numpy>=1.24.0
pyyaml>=6.0
scipy>=1.10.0
soundfile>=0.12.0
